requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
python-docx==1.0.1
pandas==2.1.3
google-generativeai==0.3.1
python-dotenv==1.0.0